Kill AppSearch.exe
Alternatives to Search:
- Everything (VoidTools)

Side-Effects: Web Search will not work (Cortana).

Why is it listed under "Outdated".
It takes only some CPU processing power during search events and runs on low-prio in background. It takes around 29-32 MB (private bytes) of RAM and is considerable not an "fps" or "system killer". However, every resource counts and you can decide yourself if you use external tools (or not) and if you like to kill some of Cortana features or not. 

The origin is more that if you like to kill DWM, you need to stop SearchApp.exe as well as StartMenuExperienceHost.exe too and then such an example script comes in handy. For more info regarding killing dwm, check out the script here:
https://github.com/Ingan121/DWMKiller

And the discussion/instructions given here:
https://winclassic.boards.net/thread/574/reliable-non-intrusive-disabling-windows?page=1&scrollTo=4613

The side-effect in killing DWM is that third-party tools like StartIsBack++ are going to break because they inject directly into specific processes which aren't started if DWM isn't running. You then need to use Windows own Start Menu or something which does not depend on it like e.g. OpenShellMenu.